add getLastMedia* methods into HasMediaTrait
It is just a small adding. I need a shortcut to get the latest image from a collection in many cases such as getting the latest uploaded profile picture of a user, while preserving the old ones for reuse.

I don't know about the performance consideration, but I think it is complementing for the _getFirst*_ pair. I use this additional snippet, hopefully it could be official :)