Use \Storage class to getMimetype
There used to be no way to retrieve the Mime type for files located on S3 since the code would return early if it wasn't a 'local' disk. With this proposed change, the `getTypeFromMimeAttribute()` method now attempts to get the mime type for any disk (local, s3, etc.) and uses the Storage class to do it.