Fix Breaking Change by using Optional Paramater
Move the `$conversionQuality` parameter to the end of `generareResponsiveImage` method and make it optional through use of the `DEFAULT_CONVERSION_QUALITY` class constant.

Define class constant, `DEFAULT_CONVERSION_QUALITY` so that this int can be defined in one place only, making future changes of it easier.

Changed `Conversion` to a required parameter for the private `getConversionQuality` method, as the method will now be **opt-in** for devs, so it can be seen as intended to be called and not called through ambigous method chaining.