Fix missing image preview URLs in Inertia Projects
When using Inertia.js with Laravel, preview_url is not being populated causing missing thumbnails in the Media Library Pro Vue components. This is because the Inertia::render() method checks if a prop is arrayable. If so, it calls the toArray method. This PR adds the toArray method for proper mapping of media objects when used with Inertia.js.