Fix: Human-readable size support-function return incorrect results
getHumanReadableSize() support function return (KB) unit for (0), For a size of 0 bytes, it's more relevant to display it as 0 B (bytes). Displaying it as 0 KB could be misleading, implying that there is a non-zero amount of data, albeit in kilobytes. Therefore, it's more appropriate to represent 0 as 0 B to accurately convey that there is no data present.