Refinement & Fix bugs (#3540)
* fix: Patch `LFI` vulnerability in `createMultipleFromRequest -> FileAdderFactory::class` method

- Implemented input sanitization to prevent directory traversal attacks
- Removed unsafe file path manipulation in createMultipleFromRequest method
- Added validation checks to ensure file paths are safe before processing

* Fix: Human-readable size support-function return incorrect results

getHumanReadableSize() support function return (KB) unit for (0), For a size of 0 bytes, it's more relevant to display it as 0 B (bytes). Displaying it as 0 KB could be misleading, implying that there is a non-zero amount of data, albeit in kilobytes. Therefore, it's more appropriate to represent 0 as 0 B to accurately convey that there is no data present.

* Fix: file name optimizer from `FileAdderFactory`

* Adding tests for larger file size support `[TB, PB, EB, ZB, YB]` - Qualified Code

* Testing negative byte value in `getHumanReadableSize()` support function

---------

Co-authored-by: TheXerr0r <efma.tech@gmail.com>